requirements.txt
pandas
numpy
openpyxl
requests
pycountry
pdfminer.six
matplotlib
scipy